requirements.txt
numpy==2.2.6
faiss-cpu==1.12.0
streamlit==1.50.0
python-dotenv==1.1.1
openai==1.109.1
PyPDF2==3.0.1
python-docx==1.2.0
